Fix crash when a random file already exists within the directory

+++ SortOfRandomFileChooser/Program.cs
@@ -83,7 +83,11 @@ namespace SortOfRandomFileChooser
                     {
                         Directory.CreateDirectory(FileManagement.RelativeToApplication("Random Files"));
                     }
-                    File.Copy(filePaths, FileManagement.RelativeToApplication("Random Files", x[x.Length - 1]));
+
+                    if (!File.Exists(FileManagement.RelativeToApplication("Random Files", x[x.Length - 1])))
+                    {
+                        File.Copy(filePaths, FileManagement.RelativeToApplication("Random Files", x[x.Length - 1]));
+                    }
                 }
 
                 Console.WriteLine("File Paths Exported to: {0}", FileManagement.RelativeToApplication("RetrievedFiles.txt"));
